Move code according to its role
