+++ frontend/src/mocks/data/users.ts
@@ -1,6 +1,6 @@
 import { GUEST_EMAIL, GUEST_PASSWORD } from 'config/app';
 import { User } from 'models/User';
-import { sanitizeUser, users, usersData, UsersSchema } from 'mocks/models/user';
+import { UsersDataType, UsersSchema } from 'mocks/models/user';
 import { digestText } from 'mocks/utils/crypto';
 import { exists, load, save } from 'mocks/utils/data';
 
@@ -22,6 +22,8 @@ export const unverifiedUser: User = {
   updatedAt: new Date(),
 };
 
+export const usersData: UsersDataType = {};
+
 const initialUsers: User[] = [guestUser, unverifiedUser];
 
 const initialize = () => {
@@ -38,7 +40,6 @@ const initialize = () => {
     const uuid = String(user.id);
 
     usersData[uuid] = userData;
-    users.push(sanitizeUser(userData));
 
     save('usersData', userData);
   });
